# use libm remquo and copysign for a little speedup