--- a/Cargo.toml
+++ b/Cargo.toml
@@ -25,6 +25,7 @@ default = ["accurate"]
 
 [dependencies]
 accurate = { version = "0.3", optional = true, default-features = false }
+libm = { version = "0.2.8", default-features = false }
 
 [dev-dependencies]
 approx = "0.5.1"